Revert "Merge branch 'main' into HUYPS39177"
This reverts commit 287a1ed688d773ce9260ebfbb5ae054188649e59, reversing
changes made to 61007cc0f5ea6b6eebcd8b5c0be8d49ca11a8ecf.

+++ AOS_BE/src/main/java/com/aos/AOSBE/Service/ReviewsService.java
@@ -54,7 +54,7 @@ public class ReviewsService {
 	}
 
 	@Transactional
-	public Double getAverageRatingByProductItemId(Integer productItemId) {
+	public Double getAverageRatingByProductItemId(Long productItemId) {
     Double avg = reviewsRepository.findAverageRatingByProductItemId(productItemId);
     return avg != null ? Math.round(avg * 10.0) / 10.0 : 0.0;
 	}
